// src/app/candidate.interface.ts
export interface Candidate {
  idcandidates?:number;
  name: string;
  surname: string;
  gender: boolean;
  dateOfBirth: Date;
  placeOfBirth: string;
  originRegion: string;
  originDivision: string;
  originSubDivision: string;
  employmentSituation: string;
  typeHandicap: string;
  nationality: string;
  cardID: string;
  cardIDIssueDate: string;
  language1: string;
  language2: string;
  maritalStatus: string;
  numberOfChildren: number;
  specialityDiploma: string;
  seriesDiploma: string;
  avgDiploma: number;
  diplomaObtainYear: string;
  degreeDiploma: string;
  obtainingInstitution: string;
  examCenter: string;
  depositPlace: string;
  optionOfTraining: string;
  admissionMode: string;
  cycle: string;
  department: string;
  option: string;
  session: string;
  residenceCity: string;
  tel1: string;
  tel2: string;
  email: string;
  bp: string;
  fax: string;
  address: string
  practicedSport: string;
  othersActivities: string;
  dateFirstRegistration: Date;
  dateModification: Date;
  dateValidation: Date;
  orderNumber: number;
  dateForDeposit: number;
  isValid: boolean;
  periodModification: number;
  //users_id;
}